requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pyperclip==1.9.0
python-owasp-zap-v2.4==0.1.0
requests==2.32.4
six==1.17.0
tk==0.1.0
urllib3==2.5.0
zaproxy==0.4.0
